moved the search part from defaultcontroller to more specific searchcontroller
