+++ alumni/src/Controller/DefaultController.php
@@ -4,10 +4,6 @@ namespace App\Controller;
 
 use Symfony\Bundle\FrameworkBundle\Controller\Controller;
 use Symfony\Component\Security\Http\Authentication\AuthenticationUtils;
-use Symfony\Component\HttpFoundation\Request;
-use App\Form\UserSearchFormType;
-use App\DTO\UserSearch;
-use App\Entity\User;
 
 
 class DefaultController extends Controller
@@ -33,27 +29,6 @@ class DefaultController extends Controller
                 'last_username' => $lastUsername,
                 'error' => $error,
             )
-        );
-    }
-
-    public function listUsers(Request $request)
-    {
-        $manager = $this->getDoctrine()->getManager();
-    
-        $dto = new UserSearch();
-        $searchForm = $this->createForm(UserSearchFormType::class, $dto, ['standalone'=>true]);
-
-        $searchForm->handleRequest($request);
-
-        $users = $manager->getRepository(User::class)->findByUserSearch($dto);
-    
-        return $this->render(
-            'Userlist/list.html.twig',
-            [
-                'users' =>  $users,
-                'searchForm'=>$searchForm->createView()
-
-            ]
         );
     }
